requirements.txt
aiogram==2.16
Django==3.2.9
libmesh==0.0.6